requirements.txt
pandas>=2.0
numpy>=1.24
yfinance>=0.2
vectorbt>=0.28
scikit-learn>=1.3
xgboost>=2.0
transformers>=4.30
requests>=2.31
pyyaml>=6.0
schedule>=1.2
loguru>=0.7
